@@ -1,53 +1,53 @@
 name: flutter_html
-description: A Flutter widget for rendering static HTML and CSS as Flutter widgets.
-version: 3.0.0-beta.2
+description: A Flutter widget rendering static HTML and CSS as Flutter widgets.
+version: 2.2.4
 homepage: https://github.com/Sub6Resources/flutter_html
+publish_to: none
 
 environment:
-  sdk: '>=2.17.0 <4.0.0'
-  flutter: '>=2.2.0'
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.2.0"
 
 dependencies:
   # Plugin for parsing html
-  html: ^0.15.3
+  html: ">=0.15.0 <1.0.0"
 
-  # Plugin for parsing css
-  csslib: ^0.17.2
+  # Plugins for parsing css
+  csslib: ^1.0.0
 
-  # Plugin for firstWhereOrNull extension on lists
-  collection: ^1.17.0
+  # Plugins for rendering the <table> tag.
+  flutter_layout_grid: ^2.0.4
 
-  # plugin to manage lists and counting in a variety of styles
-  list_counter: ^1.0.2
+  # Plugins for rendering the <video> tag.
+  video_player: ">=2.1.1 <3.0.0"
+  # chewie:
+  #   path: /Users/qonvexmacmini/Desktop/chewie
+
+  # Plugin for rendering the <iframe> tag.
+  webview_flutter: ^4.2.4
+
+  # Plugins for rendering the <audio> tag.
+  chewie_audio: ">=1.2.0 <2.0.0"
+
+  # Plugins for rendering the <svg> tag.
+  flutter_svg: ^2.0.0+1
+
+  # Plugin for rendering MathML
+  flutter_math_fork: ^0.7.1
+
+  # plugin for firstWhereOrNull extension on lists
+  collection: ">=1.15.0 <2.0.0"
+
+  # plugin to convert integers to numerals
+  numerus: ^2.1.2
 
   flutter:
     sdk: flutter
+  chewie:
+    git: https://github.com/alnaughty/chewie_custom.git
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.1
-  meta: ^1.8.0
-  melos: ^3.0.1
 
-screenshots:
-  - description: 'An example of the output produced by flutter_html'
-    path: example/screenshots/flutter_html_screenshot.png
-  - description: 'Continued example of the output produced by flutter_html'
-    path: example/screenshots/flutter_html_screenshot1.png
-  - description: 'Continued example of the output produced by flutter_html'
-    path: example/screenshots/flutter_html_screenshot2.png
-  - description: 'Continued example of the output produced by flutter_html'
-    path: example/screenshots/flutter_html_screenshot3.png
-  - description: 'Another example of the output produced by flutter_html'
-    path: example/screenshots/flutter_html_readme_screenshot.png
-
-funding:
-  - https://opencollective.com/flutter_html
-
-topics:
-  - html
-  - css
-  - widgets
-  - layout
-  - flutter_html
+flutter:
