@@ -1,5 +1,5 @@
 import 'package:flutter/widgets.dart';
-import 'package:flutter_html/src/tree/styled_element.dart';
+import 'package:flutter_html/src/styled_element.dart';
 
 class AnchorKey extends GlobalKey {
   static final Set<AnchorKey> _registry = <AnchorKey>{};
@@ -30,10 +30,7 @@ class AnchorKey extends GlobalKey {
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
-      other is AnchorKey &&
-          runtimeType == other.runtimeType &&
-          parentKey == other.parentKey &&
-          id == other.id;
+      other is AnchorKey && runtimeType == other.runtimeType && parentKey == other.parentKey && id == other.id;
 
   @override
   int get hashCode => parentKey.hashCode ^ id.hashCode;
