@@ -7,112 +7,59 @@ import 'package:flutter/material.dart';
 import 'package:flutter_html/flutter_html.dart';
 import 'package:flutter_html/src/utils.dart';
 
-//TODO refactor
 Style declarationsToStyle(Map<String, List<css.Expression>> declarations) {
-  Style style = Style();
+  Style style = new Style();
   declarations.forEach((property, value) {
     if (value.isNotEmpty) {
       switch (property) {
         case 'background-color':
-          style.backgroundColor =
-              ExpressionMapping.expressionToColor(value.first) ??
-                  style.backgroundColor;
+          style.backgroundColor = ExpressionMapping.expressionToColor(value.first) ?? style.backgroundColor;
           break;
         case 'border':
-          List<css.LiteralTerm?>? borderWidths =
-              value.whereType<css.LiteralTerm>().toList();
-
+          List<css.LiteralTerm?>? borderWidths = value.whereType<css.LiteralTerm>().toList();
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.width], so make sure to remove those before passing it to [ExpressionMapping]
-          borderWidths.removeWhere((element) =>
-              element == null ||
-              (element.text != "thin" &&
-                  element.text != "medium" &&
-                  element.text != "thick" &&
-                  element is! css.LengthTerm &&
-                  element is! css.PercentageTerm &&
-                  element is! css.EmTerm &&
-                  element is! css.RemTerm &&
-                  element is! css.NumberTerm));
-          List<css.Expression?>? borderColors = value
-              .where((element) =>
-                  ExpressionMapping.expressionToColor(element) != null)
-              .toList();
-          List<css.LiteralTerm?>? potentialStyles =
-              value.whereType<css.LiteralTerm>().toList();
-
+          borderWidths.removeWhere((element) => element == null || (element.text != "thin"
+              && element.text != "medium" && element.text != "thick"
+              && !(element is css.LengthTerm) && !(element is css.PercentageTerm)
+              && !(element is css.EmTerm) && !(element is css.RemTerm)
+              && !(element is css.NumberTerm))
+          );
+          List<css.Expression?>? borderColors = value.where((element) => ExpressionMapping.expressionToColor(element) != null).toList();
+          List<css.LiteralTerm?>? potentialStyles = value.whereType<css.LiteralTerm>().toList();
           /// Currently doesn't matter, as Flutter only supports "solid" or "none", but may support more in the future.
-          List<String> possibleBorderValues = [
-            "dotted",
-            "dashed",
-            "solid",
-            "double",
-            "groove",
-            "ridge",
-            "inset",
-            "outset",
-            "none",
-            "hidden"
-          ];
-
+          List<String> possibleBorderValues = ["dotted", "dashed", "solid", "double", "groove", "ridge", "inset", "outset", "none", "hidden"];
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.style], so make sure to remove those before passing it to [ExpressionMapping]
-          potentialStyles.removeWhere((element) =>
-              element == null || !possibleBorderValues.contains(element.text));
+          potentialStyles.removeWhere((element) => element == null || !possibleBorderValues.contains(element.text));
           List<css.LiteralTerm?>? borderStyles = potentialStyles;
-          style.border = ExpressionMapping.expressionToBorder(
-              borderWidths, borderStyles, borderColors);
+          style.border = ExpressionMapping.expressionToBorder(borderWidths, borderStyles, borderColors);
           break;
         case 'border-left':
-          List<css.LiteralTerm?>? borderWidths =
-              value.whereType<css.LiteralTerm>().toList();
-
+          List<css.LiteralTerm?>? borderWidths = value.whereType<css.LiteralTerm>().toList();
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.width], so make sure to remove those before passing it to [ExpressionMapping]
-          borderWidths.removeWhere((element) =>
-              element == null ||
-              (element.text != "thin" &&
-                  element.text != "medium" &&
-                  element.text != "thick" &&
-                  element is! css.LengthTerm &&
-                  element is! css.PercentageTerm &&
-                  element is! css.EmTerm &&
-                  element is! css.RemTerm &&
-                  element is! css.NumberTerm));
-          css.LiteralTerm? borderWidth =
-              borderWidths.firstWhereOrNull((element) => element != null);
-          css.Expression? borderColor = value.firstWhereOrNull((element) =>
-              ExpressionMapping.expressionToColor(element) != null);
-          List<css.LiteralTerm?>? potentialStyles =
-              value.whereType<css.LiteralTerm>().toList();
-
+          borderWidths.removeWhere((element) => element == null || (element.text != "thin"
+              && element.text != "medium" && element.text != "thick"
+              && !(element is css.LengthTerm) && !(element is css.PercentageTerm)
+              && !(element is css.EmTerm) && !(element is css.RemTerm)
+              && !(element is css.NumberTerm))
+          );
+          css.LiteralTerm? borderWidth = borderWidths.firstWhereOrNull((element) => element != null);
+          css.Expression? borderColor = value.firstWhereOrNull((element) => ExpressionMapping.expressionToColor(element) != null);
+          List<css.LiteralTerm?>? potentialStyles = value.whereType<css.LiteralTerm>().toList();
           /// Currently doesn't matter, as Flutter only supports "solid" or "none", but may support more in the future.
-          List<String> possibleBorderValues = [
-            "dotted",
-            "dashed",
-            "solid",
-            "double",
-            "groove",
-            "ridge",
-            "inset",
-            "outset",
-            "none",
-            "hidden"
-          ];
-
+          List<String> possibleBorderValues = ["dotted", "dashed", "solid", "double", "groove", "ridge", "inset", "outset", "none", "hidden"];
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.style], so make sure to remove those before passing it to [ExpressionMapping]
-          potentialStyles.removeWhere((element) =>
-              element == null || !possibleBorderValues.contains(element.text));
-          css.LiteralTerm? borderStyle = potentialStyles.firstOrNull;
+          potentialStyles.removeWhere((element) => element == null || !possibleBorderValues.contains(element.text));
+          css.LiteralTerm borderStyle = potentialStyles.first!;
           Border newBorder = Border(
             left: style.border?.left.copyWith(
-                  width: ExpressionMapping.expressionToBorderWidth(borderWidth),
-                  style: ExpressionMapping.expressionToBorderStyle(borderStyle),
-                  color: ExpressionMapping.expressionToColor(borderColor),
-                ) ??
-                BorderSide(
-                  width: ExpressionMapping.expressionToBorderWidth(borderWidth),
-                  style: ExpressionMapping.expressionToBorderStyle(borderStyle),
-                  color: ExpressionMapping.expressionToColor(borderColor) ??
-                      Colors.black,
-                ),
+              width: ExpressionMapping.expressionToBorderWidth(borderWidth),
+              style: ExpressionMapping.expressionToBorderStyle(borderStyle),
+              color: ExpressionMapping.expressionToColor(borderColor),
+            ) ?? BorderSide(
+              width: ExpressionMapping.expressionToBorderWidth(borderWidth),
+              style: ExpressionMapping.expressionToBorderStyle(borderStyle),
+              color: ExpressionMapping.expressionToColor(borderColor) ?? Colors.black,
+            ),
             right: style.border?.right ?? BorderSide.none,
             top: style.border?.top ?? BorderSide.none,
             bottom: style.border?.bottom ?? BorderSide.none,
@@ -120,518 +67,271 @@ Style declarationsToStyle(Map<String, List<css.Expression>> declarations) {
           style.border = newBorder;
           break;
         case 'border-right':
-          List<css.LiteralTerm?>? borderWidths =
-              value.whereType<css.LiteralTerm>().toList();
-
+          List<css.LiteralTerm?>? borderWidths = value.whereType<css.LiteralTerm>().toList();
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.width], so make sure to remove those before passing it to [ExpressionMapping]
-          borderWidths.removeWhere((element) =>
-              element == null ||
-              (element.text != "thin" &&
-                  element.text != "medium" &&
-                  element.text != "thick" &&
-                  element is! css.LengthTerm &&
-                  element is! css.PercentageTerm &&
-                  element is! css.EmTerm &&
-                  element is! css.RemTerm &&
-                  element is! css.NumberTerm));
-          css.LiteralTerm? borderWidth =
-              borderWidths.firstWhereOrNull((element) => element != null);
-          css.Expression? borderColor = value.firstWhereOrNull((element) =>
-              ExpressionMapping.expressionToColor(element) != null);
-          List<css.LiteralTerm?>? potentialStyles =
-              value.whereType<css.LiteralTerm>().toList();
-
+          borderWidths.removeWhere((element) => element == null || (element.text != "thin"
+              && element.text != "medium" && element.text != "thick"
+              && !(element is css.LengthTerm) && !(element is css.PercentageTerm)
+              && !(element is css.EmTerm) && !(element is css.RemTerm)
+              && !(element is css.NumberTerm))
+          );
+          css.LiteralTerm? borderWidth = borderWidths.firstWhereOrNull((element) => element != null);
+          css.Expression? borderColor = value.firstWhereOrNull((element) => ExpressionMapping.expressionToColor(element) != null);
+          List<css.LiteralTerm?>? potentialStyles = value.whereType<css.LiteralTerm>().toList();
           /// Currently doesn't matter, as Flutter only supports "solid" or "none", but may support more in the future.
-          List<String> possibleBorderValues = [
-            "dotted",
-            "dashed",
-            "solid",
-            "double",
-            "groove",
-            "ridge",
-            "inset",
-            "outset",
-            "none",
-            "hidden"
-          ];
-
+          List<String> possibleBorderValues = ["dotted", "dashed", "solid", "double", "groove", "ridge", "inset", "outset", "none", "hidden"];
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.style], so make sure to remove those before passing it to [ExpressionMapping]
-          potentialStyles.removeWhere((element) =>
-              element == null || !possibleBorderValues.contains(element.text));
-          css.LiteralTerm? borderStyle = potentialStyles.firstOrNull;
+          potentialStyles.removeWhere((element) => element == null || !possibleBorderValues.contains(element.text));
+          css.LiteralTerm borderStyle = potentialStyles.first!;
           Border newBorder = Border(
             left: style.border?.left ?? BorderSide.none,
             right: style.border?.right.copyWith(
-                  width: ExpressionMapping.expressionToBorderWidth(borderWidth),
-                  style: ExpressionMapping.expressionToBorderStyle(borderStyle),
-                  color: ExpressionMapping.expressionToColor(borderColor),
-                ) ??
-                BorderSide(
-                  width: ExpressionMapping.expressionToBorderWidth(borderWidth),
-                  style: ExpressionMapping.expressionToBorderStyle(borderStyle),
-                  color: ExpressionMapping.expressionToColor(borderColor) ??
-                      Colors.black,
-                ),
+              width: ExpressionMapping.expressionToBorderWidth(borderWidth),
+              style: ExpressionMapping.expressionToBorderStyle(borderStyle),
+              color: ExpressionMapping.expressionToColor(borderColor),
+            ) ?? BorderSide(
+              width: ExpressionMapping.expressionToBorderWidth(borderWidth),
+              style: ExpressionMapping.expressionToBorderStyle(borderStyle),
+              color: ExpressionMapping.expressionToColor(borderColor) ?? Colors.black,
+            ),
             top: style.border?.top ?? BorderSide.none,
             bottom: style.border?.bottom ?? BorderSide.none,
           );
           style.border = newBorder;
           break;
         case 'border-top':
-          List<css.LiteralTerm?>? borderWidths =
-              value.whereType<css.LiteralTerm>().toList();
-
+          List<css.LiteralTerm?>? borderWidths = value.whereType<css.LiteralTerm>().toList();
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.width], so make sure to remove those before passing it to [ExpressionMapping]
-          borderWidths.removeWhere((element) =>
-              element == null ||
-              (element.text != "thin" &&
-                  element.text != "medium" &&
-                  element.text != "thick" &&
-                  element is! css.LengthTerm &&
-                  element is! css.PercentageTerm &&
-                  element is! css.EmTerm &&
-                  element is! css.RemTerm &&
-                  element is! css.NumberTerm));
-          css.LiteralTerm? borderWidth =
-              borderWidths.firstWhereOrNull((element) => element != null);
-          css.Expression? borderColor = value.firstWhereOrNull((element) =>
-              ExpressionMapping.expressionToColor(element) != null);
-          List<css.LiteralTerm?>? potentialStyles =
-              value.whereType<css.LiteralTerm>().toList();
-
+          borderWidths.removeWhere((element) => element == null || (element.text != "thin"
+              && element.text != "medium" && element.text != "thick"
+              && !(element is css.LengthTerm) && !(element is css.PercentageTerm)
+              && !(element is css.EmTerm) && !(element is css.RemTerm)
+              && !(element is css.NumberTerm))
+          );
+          css.LiteralTerm? borderWidth = borderWidths.firstWhereOrNull((element) => element != null);
+          css.Expression? borderColor = value.firstWhereOrNull((element) => ExpressionMapping.expressionToColor(element) != null);
+          List<css.LiteralTerm?>? potentialStyles = value.whereType<css.LiteralTerm>().toList();
           /// Currently doesn't matter, as Flutter only supports "solid" or "none", but may support more in the future.
-          List<String> possibleBorderValues = [
-            "dotted",
-            "dashed",
-            "solid",
-            "double",
-            "groove",
-            "ridge",
-            "inset",
-            "outset",
-            "none",
-            "hidden"
-          ];
-
+          List<String> possibleBorderValues = ["dotted", "dashed", "solid", "double", "groove", "ridge", "inset", "outset", "none", "hidden"];
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.style], so make sure to remove those before passing it to [ExpressionMapping]
-          potentialStyles.removeWhere((element) =>
-              element == null || !possibleBorderValues.contains(element.text));
-          css.LiteralTerm? borderStyle = potentialStyles.firstOrNull;
+          potentialStyles.removeWhere((element) => element == null || !possibleBorderValues.contains(element.text));
+          css.LiteralTerm borderStyle = potentialStyles.first!;
           Border newBorder = Border(
             left: style.border?.left ?? BorderSide.none,
             right: style.border?.right ?? BorderSide.none,
             top: style.border?.top.copyWith(
-                  width: ExpressionMapping.expressionToBorderWidth(borderWidth),
-                  style: ExpressionMapping.expressionToBorderStyle(borderStyle),
-                  color: ExpressionMapping.expressionToColor(borderColor),
-                ) ??
-                BorderSide(
-                  width: ExpressionMapping.expressionToBorderWidth(borderWidth),
-                  style: ExpressionMapping.expressionToBorderStyle(borderStyle),
-                  color: ExpressionMapping.expressionToColor(borderColor) ??
-                      Colors.black,
-                ),
+              width: ExpressionMapping.expressionToBorderWidth(borderWidth),
+              style: ExpressionMapping.expressionToBorderStyle(borderStyle),
+              color: ExpressionMapping.expressionToColor(borderColor),
+            ) ?? BorderSide(
+              width: ExpressionMapping.expressionToBorderWidth(borderWidth),
+              style: ExpressionMapping.expressionToBorderStyle(borderStyle),
+              color: ExpressionMapping.expressionToColor(borderColor) ?? Colors.black,
+            ),
             bottom: style.border?.bottom ?? BorderSide.none,
           );
           style.border = newBorder;
           break;
         case 'border-bottom':
-          List<css.LiteralTerm?>? borderWidths =
-              value.whereType<css.LiteralTerm>().toList();
-
+          List<css.LiteralTerm?>? borderWidths = value.whereType<css.LiteralTerm>().toList();
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.width], so make sure to remove those before passing it to [ExpressionMapping]
-          borderWidths.removeWhere((element) =>
-              element == null ||
-              (element.text != "thin" &&
-                  element.text != "medium" &&
-                  element.text != "thick" &&
-                  element is! css.LengthTerm &&
-                  element is! css.PercentageTerm &&
-                  element is! css.EmTerm &&
-                  element is! css.RemTerm &&
-                  element is! css.NumberTerm));
-          css.LiteralTerm? borderWidth =
-              borderWidths.firstWhereOrNull((element) => element != null);
-          css.Expression? borderColor = value.firstWhereOrNull((element) =>
-              ExpressionMapping.expressionToColor(element) != null);
-          List<css.LiteralTerm?>? potentialStyles =
-              value.whereType<css.LiteralTerm>().toList();
-
+          borderWidths.removeWhere((element) => element == null || (element.text != "thin"
+              && element.text != "medium" && element.text != "thick"
+              && !(element is css.LengthTerm) && !(element is css.PercentageTerm)
+              && !(element is css.EmTerm) && !(element is css.RemTerm)
+              && !(element is css.NumberTerm))
+          );
+          css.LiteralTerm? borderWidth = borderWidths.firstWhereOrNull((element) => element != null);
+          css.Expression? borderColor = value.firstWhereOrNull((element) => ExpressionMapping.expressionToColor(element) != null);
+          List<css.LiteralTerm?>? potentialStyles = value.whereType<css.LiteralTerm>().toList();
           /// Currently doesn't matter, as Flutter only supports "solid" or "none", but may support more in the future.
-          List<String> possibleBorderValues = [
-            "dotted",
-            "dashed",
-            "solid",
-            "double",
-            "groove",
-            "ridge",
-            "inset",
-            "outset",
-            "none",
-            "hidden"
-          ];
-
+          List<String> possibleBorderValues = ["dotted", "dashed", "solid", "double", "groove", "ridge", "inset", "outset", "none", "hidden"];
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.style], so make sure to remove those before passing it to [ExpressionMapping]
-          potentialStyles.removeWhere((element) =>
-              element == null || !possibleBorderValues.contains(element.text));
+          potentialStyles.removeWhere((element) => element == null || !possibleBorderValues.contains(element.text));
           css.LiteralTerm? borderStyle = potentialStyles.firstOrNull;
           Border newBorder = Border(
             left: style.border?.left ?? BorderSide.none,
             right: style.border?.right ?? BorderSide.none,
             top: style.border?.top ?? BorderSide.none,
             bottom: style.border?.bottom.copyWith(
-                  width: ExpressionMapping.expressionToBorderWidth(borderWidth),
-                  style: ExpressionMapping.expressionToBorderStyle(borderStyle),
-                  color: ExpressionMapping.expressionToColor(borderColor),
-                ) ??
-                BorderSide(
-                  width: ExpressionMapping.expressionToBorderWidth(borderWidth),
-                  style: ExpressionMapping.expressionToBorderStyle(borderStyle),
-                  color: ExpressionMapping.expressionToColor(borderColor) ??
-                      Colors.black,
-                ),
+              width: ExpressionMapping.expressionToBorderWidth(borderWidth),
+              style: ExpressionMapping.expressionToBorderStyle(borderStyle),
+              color: ExpressionMapping.expressionToColor(borderColor),
+            ) ?? BorderSide(
+              width: ExpressionMapping.expressionToBorderWidth(borderWidth),
+              style: ExpressionMapping.expressionToBorderStyle(borderStyle),
+              color: ExpressionMapping.expressionToColor(borderColor) ?? Colors.black,
+            ),
           );
           style.border = newBorder;
           break;
         case 'color':
-          style.color =
-              ExpressionMapping.expressionToColor(value.first) ?? style.color;
+          style.color = ExpressionMapping.expressionToColor(value.first) ?? style.color;
           break;
         case 'direction':
-          style.direction =
-              ExpressionMapping.expressionToDirection(value.first);
+          style.direction = ExpressionMapping.expressionToDirection(value.first);
           break;
         case 'display':
           style.display = ExpressionMapping.expressionToDisplay(value.first);
           break;
         case 'line-height':
-          style.lineHeight =
-              ExpressionMapping.expressionToLineHeight(value.first);
+          style.lineHeight = ExpressionMapping.expressionToLineHeight(value.first);
           break;
         case 'font-family':
-          style.fontFamily =
-              ExpressionMapping.expressionToFontFamily(value.first) ??
-                  style.fontFamily;
+          style.fontFamily = ExpressionMapping.expressionToFontFamily(value.first) ?? style.fontFamily;
           break;
         case 'font-feature-settings':
-          style.fontFeatureSettings =
-              ExpressionMapping.expressionToFontFeatureSettings(value);
+          style.fontFeatureSettings = ExpressionMapping.expressionToFontFeatureSettings(value);
           break;
         case 'font-size':
-          style.fontSize =
-              ExpressionMapping.expressionToFontSize(value.first) ??
-                  style.fontSize;
+          style.fontSize = ExpressionMapping.expressionToFontSize(value.first) ?? style.fontSize;
           break;
         case 'font-style':
-          style.fontStyle =
-              ExpressionMapping.expressionToFontStyle(value.first);
+          style.fontStyle = ExpressionMapping.expressionToFontStyle(value.first);
           break;
         case 'font-weight':
-          style.fontWeight =
-              ExpressionMapping.expressionToFontWeight(value.first);
+          style.fontWeight = ExpressionMapping.expressionToFontWeight(value.first);
           break;
         case 'list-style':
-          css.LiteralTerm? position = value.firstWhereOrNull((e) =>
-              e is css.LiteralTerm &&
-              (e.text == "outside" || e.text == "inside")) as css.LiteralTerm?;
-          css.UriTerm? image =
-              value.firstWhereOrNull((e) => e is css.UriTerm) as css.UriTerm?;
-          css.LiteralTerm? type = value.firstWhereOrNull((e) =>
-              e is css.LiteralTerm &&
-              e.text != "outside" &&
-              e.text != "inside") as css.LiteralTerm?;
+          css.LiteralTerm? position = value.firstWhereOrNull((e) => e is css.LiteralTerm && (e.text == "outside" || e.text == "inside")) as css.LiteralTerm?;
+          css.UriTerm? image = value.firstWhereOrNull((e) => e is css.UriTerm) as css.UriTerm?;
+          css.LiteralTerm? type = value.firstWhereOrNull((e) => e is css.LiteralTerm && e.text != "outside" && e.text != "inside") as css.LiteralTerm?;
           if (position != null) {
             switch (position.text) {
               case 'outside':
-                style.listStylePosition = ListStylePosition.outside;
+                style.listStylePosition = ListStylePosition.OUTSIDE;
                 break;
               case 'inside':
-                style.listStylePosition = ListStylePosition.inside;
+                style.listStylePosition = ListStylePosition.INSIDE;
                 break;
             }
           }
           if (image != null) {
-            style.listStyleImage =
-                ExpressionMapping.expressionToListStyleImage(image) ??
-                    style.listStyleImage;
+            style.listStyleType = ExpressionMapping.expressionToListStyleType(image) ?? style.listStyleType;
           } else if (type != null) {
-            style.listStyleType =
-                ExpressionMapping.expressionToListStyleType(type) ??
-                    style.listStyleType;
+            style.listStyleType = ExpressionMapping.expressionToListStyleType(type) ?? style.listStyleType;
           }
           break;
         case 'list-style-image':
           if (value.first is css.UriTerm) {
-            style.listStyleImage = ExpressionMapping.expressionToListStyleImage(
-                    value.first as css.UriTerm) ??
-                style.listStyleImage;
+            style.listStyleType = ExpressionMapping.expressionToListStyleType(value.first as css.UriTerm) ?? style.listStyleType;
           }
           break;
         case 'list-style-position':
           if (value.first is css.LiteralTerm) {
             switch ((value.first as css.LiteralTerm).text) {
               case 'outside':
-                style.listStylePosition = ListStylePosition.outside;
+                style.listStylePosition = ListStylePosition.OUTSIDE;
                 break;
               case 'inside':
-                style.listStylePosition = ListStylePosition.inside;
+                style.listStylePosition = ListStylePosition.INSIDE;
                 break;
             }
           }
           break;
         case 'height':
-          style.height =
-              ExpressionMapping.expressionToHeight(value.first) ?? style.height;
+          style.height = ExpressionMapping.expressionToPaddingLength(value.first) ?? style.height;
           break;
         case 'list-style-type':
           if (value.first is css.LiteralTerm) {
-            style.listStyleType = ExpressionMapping.expressionToListStyleType(
-                    value.first as css.LiteralTerm) ??
-                style.listStyleType;
+            style.listStyleType = ExpressionMapping.expressionToListStyleType(value.first as css.LiteralTerm) ?? style.listStyleType;
           }
           break;
         case 'margin':
-          List<css.LiteralTerm>? marginLengths =
-              value.whereType<css.LiteralTerm>().toList();
-
+          List<css.LiteralTerm>? marginLengths = value.whereType<css.LiteralTerm>().toList();
           /// List<css.LiteralTerm> might include other values than the ones we want for margin length, so make sure to remove those before passing it to [ExpressionMapping]
-          marginLengths.removeWhere((element) =>
-              element is! css.LengthTerm &&
-              element is! css.EmTerm &&
-              element is! css.RemTerm &&
-              element is! css.NumberTerm &&
-              !(element.text == 'auto'));
-          Margins margin = ExpressionMapping.expressionToMargins(marginLengths);
-          style.margin = (style.margin ?? const Margins()).copyWith(
-            left: margin.left,
-            right: margin.right,
-            top: margin.top,
-            bottom: margin.bottom,
+          marginLengths.removeWhere((element) => !(element is css.LengthTerm)
+              && !(element is css.EmTerm)
+              && !(element is css.RemTerm)
+              && !(element is css.NumberTerm)
+          );
+          List<double?> margin = ExpressionMapping.expressionToPadding(marginLengths);
+          style.margin = (style.margin ?? EdgeInsets.zero).copyWith(
+            left: margin[0],
+            right: margin[1],
+            top: margin[2],
+            bottom: margin[3],
           );
           break;
         case 'margin-left':
-          style.margin = (style.margin ?? const Margins()).copyWith(
-              left: ExpressionMapping.expressionToMargin(value.first));
+          style.margin = (style.margin ?? EdgeInsets.zero).copyWith(
+              left: ExpressionMapping.expressionToPaddingLength(value.first));
           break;
         case 'margin-right':
-          style.margin = (style.margin ?? const Margins()).copyWith(
-              right: ExpressionMapping.expressionToMargin(value.first));
+          style.margin = (style.margin ?? EdgeInsets.zero).copyWith(
+              right: ExpressionMapping.expressionToPaddingLength(value.first));
           break;
         case 'margin-top':
-          style.margin = (style.margin ?? const Margins())
-              .copyWith(top: ExpressionMapping.expressionToMargin(value.first));
+          style.margin = (style.margin ?? EdgeInsets.zero).copyWith(
+              top: ExpressionMapping.expressionToPaddingLength(value.first));
           break;
         case 'margin-bottom':
-          style.margin = (style.margin ?? const Margins()).copyWith(
-              bottom: ExpressionMapping.expressionToMargin(value.first));
-          break;
-        case 'margin-inline':
-          List<css.LiteralTerm>? marginLengths =
-              value.whereType<css.LiteralTerm>().toList();
-
-          /// List<css.LiteralTerm> might include other values than the ones we want for margin length, so make sure to remove those before passing it to [ExpressionMapping]
-          marginLengths.removeWhere((element) =>
-              element is! css.LengthTerm &&
-              element is! css.EmTerm &&
-              element is! css.RemTerm &&
-              element is! css.NumberTerm &&
-              !(element.text == 'auto'));
-          Margins margin =
-              ExpressionMapping.expressionToInlineMargins(marginLengths);
-          style.margin = (style.margin ?? const Margins()).copyWith(
-            inlineStart: margin.inlineStart,
-            inlineEnd: margin.inlineEnd,
-          );
-          break;
-        case 'margin-inline-end':
-          style.margin = (style.margin ?? const Margins()).copyWith(
-            inlineEnd: ExpressionMapping.expressionToMargin(value.first),
-          );
-          break;
-        case 'margin-inline-start':
-          style.margin = (style.margin ?? const Margins()).copyWith(
-            inlineStart: ExpressionMapping.expressionToMargin(value.first),
-          );
-          break;
-        case 'margin-block':
-          List<css.LiteralTerm>? marginLengths =
-              value.whereType<css.LiteralTerm>().toList();
-
-          /// List<css.LiteralTerm> might include other values than the ones we want for margin length, so make sure to remove those before passing it to [ExpressionMapping]
-          marginLengths.removeWhere((element) =>
-              element is! css.LengthTerm &&
-              element is! css.EmTerm &&
-              element is! css.RemTerm &&
-              element is! css.NumberTerm &&
-              !(element.text == 'auto'));
-          Margins margin =
-              ExpressionMapping.expressionToBlockMargins(marginLengths);
-          style.margin = (style.margin ?? const Margins()).copyWith(
-            blockStart: margin.blockStart,
-            blockEnd: margin.blockEnd,
-          );
-          break;
-        case 'margin-block-end':
-          style.margin = (style.margin ?? const Margins()).copyWith(
-            blockEnd: ExpressionMapping.expressionToMargin(value.first),
-          );
-          break;
-        case 'margin-block-start':
-          style.margin = (style.margin ?? const Margins()).copyWith(
-            blockStart: ExpressionMapping.expressionToMargin(value.first),
-          );
+          style.margin = (style.margin ?? EdgeInsets.zero).copyWith(
+              bottom: ExpressionMapping.expressionToPaddingLength(value.first));
           break;
         case 'padding':
-          List<css.LiteralTerm>? paddingLengths =
-              value.whereType<css.LiteralTerm>().toList();
-
+          List<css.LiteralTerm>? paddingLengths = value.whereType<css.LiteralTerm>().toList();
           /// List<css.LiteralTerm> might include other values than the ones we want for padding length, so make sure to remove those before passing it to [ExpressionMapping]
-          paddingLengths.removeWhere((element) =>
-              element is! css.LengthTerm &&
-              element is! css.EmTerm &&
-              element is! css.RemTerm &&
-              element is! css.NumberTerm);
-          final padding =
-              ExpressionMapping.expressionToHtmlPaddings(paddingLengths);
-          style.padding = (style.padding ?? const HtmlPaddings()).copyWith(
-            left: padding.left,
-            right: padding.right,
-            top: padding.top,
-            bottom: padding.bottom,
+          paddingLengths.removeWhere((element) => !(element is css.LengthTerm)
+              && !(element is css.EmTerm)
+              && !(element is css.RemTerm)
+              && !(element is css.NumberTerm)
+          );
+          List<double?> padding = ExpressionMapping.expressionToPadding(paddingLengths);
+          style.padding = (style.padding ?? EdgeInsets.zero).copyWith(
+            left: padding[0],
+            right: padding[1],
+            top: padding[2],
+            bottom: padding[3],
           );
           break;
         case 'padding-left':
-          style.padding = (style.padding ?? const HtmlPaddings()).copyWith(
-              left: ExpressionMapping.expressionToHtmlPadding(value.first));
+          style.padding = (style.padding ?? EdgeInsets.zero).copyWith(
+              left: ExpressionMapping.expressionToPaddingLength(value.first));
           break;
         case 'padding-right':
-          style.padding = (style.padding ?? const HtmlPaddings()).copyWith(
-              right: ExpressionMapping.expressionToHtmlPadding(value.first));
+          style.padding = (style.padding ?? EdgeInsets.zero).copyWith(
+              right: ExpressionMapping.expressionToPaddingLength(value.first));
           break;
         case 'padding-top':
-          style.padding = (style.padding ?? const HtmlPaddings()).copyWith(
-              top: ExpressionMapping.expressionToHtmlPadding(value.first));
+          style.padding = (style.padding ?? EdgeInsets.zero).copyWith(
+              top: ExpressionMapping.expressionToPaddingLength(value.first));
           break;
         case 'padding-bottom':
-          style.padding = (style.padding ?? const HtmlPaddings()).copyWith(
-              bottom: ExpressionMapping.expressionToHtmlPadding(value.first));
-          break;
-        case 'padding-inline':
-          List<css.LiteralTerm>? paddingLengths =
-              value.whereType<css.LiteralTerm>().toList();
-
-          /// List<css.LiteralTerm> might include other values than the ones we want for padding length, so make sure to remove those before passing it to [ExpressionMapping]
-          paddingLengths.removeWhere((element) =>
-              element is! css.LengthTerm &&
-              element is! css.EmTerm &&
-              element is! css.RemTerm &&
-              element is! css.NumberTerm);
-          HtmlPaddings padding =
-              ExpressionMapping.expressionToInlineHtmlPadding(paddingLengths);
-          style.padding = (style.padding ?? const HtmlPaddings()).copyWith(
-            inlineStart: padding.inlineStart,
-            inlineEnd: padding.inlineEnd,
-          );
-          break;
-        case 'padding-inline-end':
-          style.padding = (style.padding ?? const HtmlPaddings()).copyWith(
-            inlineEnd: ExpressionMapping.expressionToHtmlPadding(value.first),
-          );
-          break;
-        case 'padding-inline-start':
-          style.padding = (style.padding ?? const HtmlPaddings()).copyWith(
-            inlineStart: ExpressionMapping.expressionToHtmlPadding(value.first),
-          );
-          break;
-        case 'padding-block':
-          List<css.LiteralTerm>? paddingLengths =
-              value.whereType<css.LiteralTerm>().toList();
-
-          /// List<css.LiteralTerm> might include other values than the ones we want for padding length, so make sure to remove those before passing it to [ExpressionMapping]
-          paddingLengths.removeWhere((element) =>
-              element is! css.LengthTerm &&
-              element is! css.EmTerm &&
-              element is! css.RemTerm &&
-              element is! css.NumberTerm);
-          HtmlPaddings padding =
-              ExpressionMapping.expressionToBlockHtmlPadding(paddingLengths);
-          style.padding = (style.padding ?? const HtmlPaddings()).copyWith(
-            blockStart: padding.blockStart,
-            blockEnd: padding.blockEnd,
-          );
-          break;
-        case 'padding-block-end':
-          style.padding = (style.padding ?? const HtmlPaddings()).copyWith(
-            blockEnd: ExpressionMapping.expressionToHtmlPadding(value.first),
-          );
-          break;
-        case 'padding-block-start':
-          style.padding = (style.padding ?? const HtmlPaddings()).copyWith(
-            blockStart: ExpressionMapping.expressionToHtmlPadding(value.first),
-          );
+          style.padding = (style.padding ?? EdgeInsets.zero).copyWith(
+              bottom: ExpressionMapping.expressionToPaddingLength(value.first));
           break;
         case 'text-align':
-          style.textAlign =
-              ExpressionMapping.expressionToTextAlign(value.first);
+          style.textAlign = ExpressionMapping.expressionToTextAlign(value.first);
           break;
         case 'text-decoration':
-          List<css.LiteralTerm?>? textDecorationList =
-              value.whereType<css.LiteralTerm>().toList();
-
+          List<css.LiteralTerm?>? textDecorationList = value.whereType<css.LiteralTerm>().toList();
           /// List<css.LiteralTerm> might include other values than the ones we want for [textDecorationList], so make sure to remove those before passing it to [ExpressionMapping]
-          textDecorationList.removeWhere((element) =>
-              element == null ||
-              (element.text != "none" &&
-                  element.text != "overline" &&
-                  element.text != "underline" &&
-                  element.text != "line-through"));
+          textDecorationList.removeWhere((element) => element == null || (element.text != "none"
+              && element.text != "overline" && element.text != "underline" && element.text != "line-through"));
           List<css.Expression?>? nullableList = value;
           css.Expression? textDecorationColor;
-          textDecorationColor = nullableList.firstWhereOrNull((element) =>
-              element is css.HexColorTerm || element is css.FunctionTerm);
-          List<css.LiteralTerm?>? potentialStyles =
-              value.whereType<css.LiteralTerm>().toList();
-
+          textDecorationColor = nullableList.firstWhereOrNull(
+                  (element) => element is css.HexColorTerm || element is css.FunctionTerm);
+          List<css.LiteralTerm?>? potentialStyles = value.whereType<css.LiteralTerm>().toList();
           /// List<css.LiteralTerm> might include other values than the ones we want for [textDecorationStyle], so make sure to remove those before passing it to [ExpressionMapping]
-          potentialStyles.removeWhere((element) =>
-              element == null ||
-              (element.text != "solid" &&
-                  element.text != "double" &&
-                  element.text != "dashed" &&
-                  element.text != "dotted" &&
-                  element.text != "wavy"));
-          css.LiteralTerm? textDecorationStyle =
-              potentialStyles.isNotEmpty ? potentialStyles.last : null;
-          style.textDecoration =
-              ExpressionMapping.expressionToTextDecorationLine(
-                  textDecorationList);
-          if (textDecorationColor != null) {
-            style.textDecorationColor =
-                ExpressionMapping.expressionToColor(textDecorationColor) ??
-                    style.textDecorationColor;
-          }
-          if (textDecorationStyle != null) {
-            style.textDecorationStyle =
-                ExpressionMapping.expressionToTextDecorationStyle(
-                    textDecorationStyle);
-          }
+          potentialStyles.removeWhere((element) => element == null || (element.text != "solid"
+              && element.text != "double" && element.text != "dashed" && element.text != "dotted" && element.text != "wavy"));
+          css.LiteralTerm? textDecorationStyle = potentialStyles.isNotEmpty ? potentialStyles.last : null;
+          style.textDecoration = ExpressionMapping.expressionToTextDecorationLine(textDecorationList);
+          if (textDecorationColor != null) style.textDecorationColor = ExpressionMapping.expressionToColor(textDecorationColor)
+              ?? style.textDecorationColor;
+          if (textDecorationStyle != null) style.textDecorationStyle = ExpressionMapping.expressionToTextDecorationStyle(textDecorationStyle);
           break;
         case 'text-decoration-color':
-          style.textDecorationColor =
-              ExpressionMapping.expressionToColor(value.first) ??
-                  style.textDecorationColor;
+          style.textDecorationColor = ExpressionMapping.expressionToColor(value.first) ?? style.textDecorationColor;
           break;
         case 'text-decoration-line':
-          List<css.LiteralTerm?>? textDecorationList =
-              value.whereType<css.LiteralTerm>().toList();
-          style.textDecoration =
-              ExpressionMapping.expressionToTextDecorationLine(
-                  textDecorationList);
+          List<css.LiteralTerm?>? textDecorationList = value.whereType<css.LiteralTerm>().toList();
+          style.textDecoration = ExpressionMapping.expressionToTextDecorationLine(textDecorationList);
           break;
         case 'text-decoration-style':
-          style.textDecorationStyle =
-              ExpressionMapping.expressionToTextDecorationStyle(
-                  value.first as css.LiteralTerm);
+          style.textDecorationStyle = ExpressionMapping.expressionToTextDecorationStyle(value.first as css.LiteralTerm);
           break;
         case 'text-shadow':
           style.textShadow = ExpressionMapping.expressionToTextShadow(value);
@@ -648,13 +348,8 @@ Style declarationsToStyle(Map<String, List<css.Expression>> declarations) {
             style.textTransform = TextTransform.none;
           }
           break;
-        case 'vertical-align':
-          style.verticalAlign =
-              ExpressionMapping.expressionToVerticalAlign(value.first);
-          break;
         case 'width':
-          style.width =
-              ExpressionMapping.expressionToWidth(value.first) ?? style.width;
+          style.width = ExpressionMapping.expressionToPaddingLength(value.first) ?? style.width;
           break;
       }
     }
@@ -662,7 +357,6 @@ Style declarationsToStyle(Map<String, List<css.Expression>> declarations) {
   return style;
 }
 
-//TODO refactor
 Style? inlineCssToStyle(String? inlineStyle, OnCssParseError? errorHandler) {
   var errors = <cssparser.Message>[];
   final sheet = cssparser.parse("*{$inlineStyle}", errors: errors);
@@ -678,11 +372,7 @@ Style? inlineCssToStyle(String? inlineStyle, OnCssParseError? errorHandler) {
   return null;
 }
 
-//TODO refactor?
-Map<String, Map<String, List<css.Expression>>> parseExternalCss(
-    String css, OnCssParseError? errorHandler) {
-  if (css.isEmpty) return {};
-
+Map<String, Map<String, List<css.Expression>>> parseExternalCss(String css, OnCssParseError? errorHandler) {
   var errors = <cssparser.Message>[];
   final sheet = cssparser.parse(css, errors: errors);
   if (errors.isEmpty) {
@@ -697,33 +387,30 @@ Map<String, Map<String, List<css.Expression>>> parseExternalCss(
 }
 
 class DeclarationVisitor extends css.Visitor {
-  final Map<String, Map<String, List<css.Expression>>> _result = {};
-  final Map<String, List<css.Expression>> _properties = {};
+  Map<String, Map<String, List<css.Expression>>> _result = {};
+  Map<String, List<css.Expression>> _properties = {};
   late String _selector;
   late String _currentProperty;
 
-  Map<String, Map<String, List<css.Expression>>> getDeclarations(
-      css.StyleSheet sheet) {
-    for (var element in sheet.topLevels) {
+  Map<String, Map<String, List<css.Expression>>> getDeclarations(css.StyleSheet sheet) {
+    sheet.topLevels.forEach((element) {
       if (element.span != null) {
         _selector = element.span!.text;
         element.visit(this);
         if (_result[_selector] != null) {
           _properties.forEach((key, value) {
             if (_result[_selector]![key] != null) {
-              _result[_selector]![key]!
-                  .addAll(List<css.Expression>.from(value));
+              _result[_selector]![key]!.addAll(new List<css.Expression>.from(value));
             } else {
-              _result[_selector]![key] = List<css.Expression>.from(value);
+              _result[_selector]![key] = new List<css.Expression>.from(value);
             }
           });
         } else {
-          _result[_selector] =
-              Map<String, List<css.Expression>>.from(_properties);
+          _result[_selector] = new Map<String, List<css.Expression>>.from(_properties);
         }
         _properties.clear();
       }
-    }
+    });
     return _result;
   }
 
@@ -746,10 +433,8 @@ class DeclarationVisitor extends css.Visitor {
 
 //Mapping functions
 class ExpressionMapping {
-  static Border expressionToBorder(
-      List<css.Expression?>? borderWidths,
-      List<css.LiteralTerm?>? borderStyles,
-      List<css.Expression?>? borderColors) {
+
+  static Border expressionToBorder(List<css.Expression?>? borderWidths, List<css.LiteralTerm?>? borderStyles, List<css.Expression?>? borderColors) {
     CustomBorderSide left = CustomBorderSide();
     CustomBorderSide top = CustomBorderSide();
     CustomBorderSide right = CustomBorderSide();
@@ -824,22 +509,11 @@ class ExpressionMapping {
       }
     }
     return Border(
-        top: BorderSide(
-            width: top.width,
-            color: top.color ?? Colors.black,
-            style: top.style),
-        right: BorderSide(
-            width: right.width,
-            color: right.color ?? Colors.black,
-            style: right.style),
-        bottom: BorderSide(
-            width: bottom.width,
-            color: bottom.color ?? Colors.black,
-            style: bottom.style),
-        left: BorderSide(
-            width: left.width,
-            color: left.color ?? Colors.black,
-            style: left.style));
+        top: BorderSide(width: top.width, color: top.color ?? Colors.black, style: top.style),
+        right: BorderSide(width: right.width, color: right.color ?? Colors.black, style: right.style),
+        bottom: BorderSide(width: bottom.width, color: bottom.color ?? Colors.black, style: bottom.style),
+        left: BorderSide(width: left.width, color: left.color ?? Colors.black, style: left.style)
+    );
   }
 
   static double expressionToBorderWidth(css.Expression? value) {
@@ -852,9 +526,7 @@ class ExpressionMapping {
     } else if (value is css.RemTerm) {
       return double.tryParse(value.text) ?? 1.0;
     } else if (value is css.LengthTerm) {
-      return double.tryParse(
-              value.text.replaceAll(RegExp(r'\s+(\d+\.\d+)\s+'), '')) ??
-          1.0;
+      return double.tryParse(value.text.replaceAll(new RegExp(r'\s+(\d+\.\d+)\s+'), '')) ?? 1.0;
     } else if (value is css.LiteralTerm) {
       switch (value.text) {
         case "thin":
@@ -894,7 +566,7 @@ class ExpressionMapping {
 
   static TextDirection expressionToDirection(css.Expression value) {
     if (value is css.LiteralTerm) {
-      switch (value.text) {
+      switch(value.text) {
         case "ltr":
           return TextDirection.ltr;
         case "rtl":
@@ -906,41 +578,32 @@ class ExpressionMapping {
 
   static Display expressionToDisplay(css.Expression value) {
     if (value is css.LiteralTerm) {
-      switch (value.text) {
+      switch(value.text) {
         case 'block':
-          return Display.block;
+          return Display.BLOCK;
         case 'inline-block':
-          return Display.inlineBlock;
+          return Display.INLINE_BLOCK;
         case 'inline':
-          return Display.inline;
+          return Display.INLINE;
         case 'list-item':
-          return Display.listItem;
+          return Display.LIST_ITEM;
         case 'none':
-          return Display.none;
+          return Display.NONE;
       }
     }
-    return Display.inline;
+    return Display.INLINE;
   }
 
-  static List<FontFeature> expressionToFontFeatureSettings(
-      List<css.Expression> value) {
+  static List<FontFeature> expressionToFontFeatureSettings(List<css.Expression> value) {
     List<FontFeature> fontFeatures = [];
     for (int i = 0; i < value.length; i++) {
       css.Expression exp = value[i];
       if (exp is css.LiteralTerm) {
-        if (exp.text != "on" &&
-            exp.text != "off" &&
-            exp.text != "1" &&
-            exp.text != "0") {
+        if (exp.text != "on" && exp.text != "off" && exp.text != "1" && exp.text != "0") {
           if (i < value.length - 1) {
-            css.Expression nextExp = value[i + 1];
-            if (nextExp is css.LiteralTerm &&
-                (nextExp.text == "on" ||
-                    nextExp.text == "off" ||
-                    nextExp.text == "1" ||
-                    nextExp.text == "0")) {
-              fontFeatures.add(FontFeature(exp.text,
-                  nextExp.text == "on" || nextExp.text == "1" ? 1 : 0));
+            css.Expression nextExp = value[i+1];
+            if (nextExp is css.LiteralTerm && (nextExp.text == "on" || nextExp.text == "off" || nextExp.text == "1" || nextExp.text == "0")) {
+              fontFeatures.add(FontFeature(exp.text, nextExp.text == "on" || nextExp.text == "1" ? 1 : 0));
             } else {
               fontFeatures.add(FontFeature.enable(exp.text));
             }
@@ -956,17 +619,15 @@ class ExpressionMapping {
 
   static FontSize? expressionToFontSize(css.Expression value) {
     if (value is css.NumberTerm) {
-      return FontSize(double.tryParse(value.text) ?? 16, Unit.px);
+      return FontSize(double.tryParse(value.text));
     } else if (value is css.PercentageTerm) {
-      return FontSize(double.tryParse(value.text) ?? 100, Unit.percent);
+      return FontSize.percent(int.tryParse(value.text)!);
     } else if (value is css.EmTerm) {
-      return FontSize(double.tryParse(value.text) ?? 1, Unit.em);
-      // } else if (value is css.RemTerm) { TODO
-      //   return FontSize.rem(double.tryParse(value.text) ?? 1, Unit.em);
+      return FontSize.em(double.tryParse(value.text));
+    } else if (value is css.RemTerm) {
+      return FontSize.rem(double.tryParse(value.text)!);
     } else if (value is css.LengthTerm) {
-      return FontSize(double.tryParse(
-              value.text.replaceAll(RegExp(r'\s+(\d+\.\d+)\s+'), '')) ??
-          16);
+      return FontSize(double.tryParse(value.text.replaceAll(new RegExp(r'\s+(\d+\.\d+)\s+'), '')));
     } else if (value is css.LiteralTerm) {
       switch (value.text) {
         case "xx-small":
@@ -990,7 +651,7 @@ class ExpressionMapping {
 
   static FontStyle expressionToFontStyle(css.Expression value) {
     if (value is css.LiteralTerm) {
-      switch (value.text) {
+      switch(value.text) {
         case "italic":
         case "oblique":
           return FontStyle.italic;
@@ -1023,7 +684,7 @@ class ExpressionMapping {
           return FontWeight.w900;
       }
     } else if (value is css.LiteralTerm) {
-      switch (value.text) {
+      switch(value.text) {
         case "bold":
           return FontWeight.bold;
         case "bolder":
@@ -1051,201 +712,87 @@ class ExpressionMapping {
     } else if (value is css.RemTerm) {
       return LineHeight.rem(double.tryParse(value.text)!);
     } else if (value is css.LengthTerm) {
-      return LineHeight(
-          double.tryParse(
-              value.text.replaceAll(RegExp(r'\s+(\d+\.\d+)\s+'), '')),
-          units: "length");
+      return LineHeight(double.tryParse(value.text.replaceAll(new RegExp(r'\s+(\d+\.\d+)\s+'), '')), units: "length");
     }
     return LineHeight.normal;
   }
 
-  static ListStyleImage? expressionToListStyleImage(css.UriTerm value) {
-    return ListStyleImage(value.text);
-  }
-
   static ListStyleType? expressionToListStyleType(css.LiteralTerm value) {
-    return ListStyleType.fromName(value.text);
-  }
-
-  static Width? expressionToWidth(css.Expression value) {
-    if ((value is css.LiteralTerm) && value.text == 'auto') {
-      return Width.auto();
-    } else {
-      final computedValue = expressionToLengthOrPercent(value);
-      return Width(computedValue.value, computedValue.unit);
+    if (value is css.UriTerm) {
+      return ListStyleType.fromImage(value.text);
     }
-  }
-
-  static Height? expressionToHeight(css.Expression value) {
-    if ((value is css.LiteralTerm) && value.text == 'auto') {
-      return Height.auto();
-    } else {
-      final computedValue = expressionToLengthOrPercent(value);
-      return Height(computedValue.value, computedValue.unit);
+    switch (value.text) {
+      case 'disc':
+        return ListStyleType.DISC;
+      case 'circle':
+        return ListStyleType.CIRCLE;
+      case 'decimal':
+        return ListStyleType.DECIMAL;
+      case 'lower-alpha':
+        return ListStyleType.LOWER_ALPHA;
+      case 'lower-latin':
+        return ListStyleType.LOWER_LATIN;
+      case 'lower-roman':
+        return ListStyleType.LOWER_ROMAN;
+      case 'square':
+        return ListStyleType.SQUARE;
+      case 'upper-alpha':
+        return ListStyleType.UPPER_ALPHA;
+      case 'upper-latin':
+        return ListStyleType.UPPER_LATIN;
+      case 'upper-roman':
+        return ListStyleType.UPPER_ROMAN;
     }
+    return null;
   }
 
-  static Margin? expressionToMargin(css.Expression value) {
-    if ((value is css.LiteralTerm) && value.text == 'auto') {
-      return Margin.auto();
-    } else {
-      final computedValue = expressionToLengthOrPercent(value);
-      return Margin(computedValue.value, computedValue.unit);
-    }
-  }
-
-  static Margins expressionToInlineMargins(List<css.Expression>? lengths) {
-    Margin? inlineStart;
-    Margin? inlineEnd;
-
+  static List<double?> expressionToPadding(List<css.Expression>? lengths) {
+    double? left;
+    double? right;
+    double? top;
+    double? bottom;
     if (lengths != null && lengths.isNotEmpty) {
-      inlineStart = expressionToMargin(lengths.first);
-      inlineEnd = expressionToMargin(lengths.last);
-    }
-
-    return Margins(inlineStart: inlineStart, inlineEnd: inlineEnd);
-  }
-
-  static Margins expressionToBlockMargins(List<css.Expression>? lengths) {
-    Margin? blockStart;
-    Margin? blockEnd;
-
-    if (lengths != null && lengths.isNotEmpty) {
-      blockStart = expressionToMargin(lengths.first);
-      blockEnd = expressionToMargin(lengths.last);
-    }
-
-    return Margins(blockStart: blockStart, blockEnd: blockEnd);
-  }
-
-  static Margins expressionToMargins(List<css.Expression>? lengths) {
-    Margin? left;
-    Margin? right;
-    Margin? top;
-    Margin? bottom;
-    if (lengths != null && lengths.isNotEmpty) {
-      top = expressionToMargin(lengths.first);
+      top = expressionToPaddingLength(lengths.first);
       if (lengths.length == 4) {
-        right = expressionToMargin(lengths[1]);
-        bottom = expressionToMargin(lengths[2]);
-        left = expressionToMargin(lengths.last);
+        right = expressionToPaddingLength(lengths[1]);
+        bottom = expressionToPaddingLength(lengths[2]);
+        left = expressionToPaddingLength(lengths.last);
       }
       if (lengths.length == 3) {
-        left = expressionToMargin(lengths[1]);
-        right = expressionToMargin(lengths[1]);
-        bottom = expressionToMargin(lengths.last);
+        left = expressionToPaddingLength(lengths[1]);
+        right = expressionToPaddingLength(lengths[1]);
+        bottom = expressionToPaddingLength(lengths.last);
       }
       if (lengths.length == 2) {
-        bottom = expressionToMargin(lengths.first);
-        left = expressionToMargin(lengths.last);
-        right = expressionToMargin(lengths.last);
+        bottom = expressionToPaddingLength(lengths.first);
+        left = expressionToPaddingLength(lengths.last);
+        right = expressionToPaddingLength(lengths.last);
       }
       if (lengths.length == 1) {
-        bottom = expressionToMargin(lengths.first);
-        left = expressionToMargin(lengths.first);
-        right = expressionToMargin(lengths.first);
+        bottom = expressionToPaddingLength(lengths.first);
+        left = expressionToPaddingLength(lengths.first);
+        right = expressionToPaddingLength(lengths.first);
       }
     }
-    return Margins(left: left, right: right, top: top, bottom: bottom);
+    return [left, right, top, bottom];
   }
 
-  static HtmlPadding? expressionToHtmlPadding(css.Expression value) {
-    final computedValue = expressionToLengthOrPercent(value);
-    return HtmlPadding(computedValue.value, computedValue.unit);
-  }
-
-  static HtmlPaddings expressionToInlineHtmlPadding(
-      List<css.Expression>? lengths) {
-    HtmlPadding? inlineStart;
-    HtmlPadding? inlineEnd;
-
-    if (lengths != null && lengths.isNotEmpty) {
-      inlineStart = expressionToHtmlPadding(lengths.first);
-      inlineEnd = expressionToHtmlPadding(lengths.last);
-    }
-
-    return HtmlPaddings(inlineStart: inlineStart, inlineEnd: inlineEnd);
-  }
-
-  static HtmlPaddings expressionToBlockHtmlPadding(
-      List<css.Expression>? lengths) {
-    HtmlPadding? blockStart;
-    HtmlPadding? blockEnd;
-
-    if (lengths != null && lengths.isNotEmpty) {
-      blockStart = expressionToHtmlPadding(lengths.first);
-      blockEnd = expressionToHtmlPadding(lengths.last);
-    }
-
-    return HtmlPaddings(blockStart: blockStart, blockEnd: blockEnd);
-  }
-
-  static HtmlPaddings expressionToHtmlPaddings(List<css.Expression>? lengths) {
-    HtmlPadding? left;
-    HtmlPadding? right;
-    HtmlPadding? top;
-    HtmlPadding? bottom;
-
-    if (lengths != null && lengths.isNotEmpty) {
-      top = expressionToHtmlPadding(lengths.first);
-      if (lengths.length == 4) {
-        right = expressionToHtmlPadding(lengths[1]);
-        bottom = expressionToHtmlPadding(lengths[2]);
-        left = expressionToHtmlPadding(lengths.last);
-      }
-      if (lengths.length == 3) {
-        left = expressionToHtmlPadding(lengths[1]);
-        right = expressionToHtmlPadding(lengths[1]);
-        bottom = expressionToHtmlPadding(lengths.last);
-      }
-      if (lengths.length == 2) {
-        bottom = expressionToHtmlPadding(lengths.first);
-        left = expressionToHtmlPadding(lengths.last);
-        right = expressionToHtmlPadding(lengths.last);
-      }
-      if (lengths.length == 1) {
-        bottom = expressionToHtmlPadding(lengths.first);
-        left = expressionToHtmlPadding(lengths.first);
-        right = expressionToHtmlPadding(lengths.first);
-      }
-    }
-    return HtmlPaddings(
-      left: left,
-      right: right,
-      top: top,
-      bottom: bottom,
-    );
-  }
-
-  static LengthOrPercent expressionToLengthOrPercent(css.Expression value) {
+  static double? expressionToPaddingLength(css.Expression value) {
     if (value is css.NumberTerm) {
-      return LengthOrPercent(double.parse(value.text));
+      return double.tryParse(value.text);
     } else if (value is css.EmTerm) {
-      return LengthOrPercent(double.parse(value.text), Unit.em);
+      return double.tryParse(value.text);
     } else if (value is css.RemTerm) {
-      return LengthOrPercent(double.parse(value.text), Unit.rem);
-      // TODO there are several other available terms processed by the CSS parser
+      return double.tryParse(value.text);
     } else if (value is css.LengthTerm) {
-      double number =
-          double.parse(value.text.replaceAll(RegExp(r'\s+(\d+\.\d+)\s+'), ''));
-      Unit unit = _unitMap(value.unit);
-      return LengthOrPercent(number, unit);
+      return double.tryParse(value.text.replaceAll(new RegExp(r'\s+(\d+\.\d+)\s+'), ''));
     }
-
-    //Ignore un-parsable input
-    return LengthOrPercent(0);
-  }
-
-  static Unit _unitMap(int cssParserUnitToken) {
-    switch (cssParserUnitToken) {
-      default:
-        return Unit.px;
-    }
+    return null;
   }
 
   static TextAlign expressionToTextAlign(css.Expression value) {
     if (value is css.LiteralTerm) {
-      switch (value.text) {
+      switch(value.text) {
         case "center":
           return TextAlign.center;
         case "left":
@@ -1263,12 +810,11 @@ class ExpressionMapping {
     return TextAlign.start;
   }
 
-  static TextDecoration expressionToTextDecorationLine(
-      List<css.LiteralTerm?> value) {
+  static TextDecoration expressionToTextDecorationLine(List<css.LiteralTerm?> value) {
     List<TextDecoration> decorationList = [];
     for (css.LiteralTerm? term in value) {
       if (term != null) {
-        switch (term.text) {
+        switch(term.text) {
           case "overline":
             decorationList.add(TextDecoration.overline);
             break;
@@ -1284,15 +830,12 @@ class ExpressionMapping {
         }
       }
     }
-    if (decorationList.contains(TextDecoration.none)) {
-      decorationList = [TextDecoration.none];
-    }
+    if (decorationList.contains(TextDecoration.none)) decorationList = [TextDecoration.none];
     return TextDecoration.combine(decorationList);
   }
 
-  static TextDecorationStyle expressionToTextDecorationStyle(
-      css.LiteralTerm value) {
-    switch (value.text) {
+  static TextDecorationStyle expressionToTextDecorationStyle(css.LiteralTerm value) {
+    switch(value.text) {
       case "wavy":
         return TextDecorationStyle.wavy;
       case "dotted":
@@ -1325,10 +868,10 @@ class ExpressionMapping {
       css.Expression? offsetX;
       css.Expression? offsetY;
       css.Expression? blurRadius;
-      css.Expression? color;
+      css.HexColorTerm? color;
       int expressionIndex = 0;
-      for (var element in list) {
-        if (element is css.HexColorTerm || element is css.FunctionTerm) {
+      list.forEach((element) {
+        if (element is css.HexColorTerm) {
           color = element;
         } else if (expressionIndex == 0) {
           offsetX = element;
@@ -1339,32 +882,23 @@ class ExpressionMapping {
         } else {
           blurRadius = element;
         }
-      }
+      });
       RegExp nonNumberRegex = RegExp(r'\s+(\d+\.\d+)\s+');
       if (offsetX is css.LiteralTerm && offsetY is css.LiteralTerm) {
-        if (color != null &&
-            ExpressionMapping.expressionToColor(color) != null) {
+        if (color != null && ExpressionMapping.expressionToColor(color) != null) {
           shadow.add(Shadow(
-            color: expressionToColor(color)!,
-            offset: Offset(
-                double.tryParse((offsetX).text.replaceAll(nonNumberRegex, ''))!,
-                double.tryParse(
-                    (offsetY).text.replaceAll(nonNumberRegex, ''))!),
-            blurRadius: (blurRadius is css.LiteralTerm)
-                ? double.tryParse(
-                    (blurRadius).text.replaceAll(nonNumberRegex, ''))!
-                : 0.0,
+              color: expressionToColor(color)!,
+              offset: Offset(
+                  double.tryParse((offsetX as css.LiteralTerm).text.replaceAll(nonNumberRegex, ''))!,
+                  double.tryParse((offsetY as css.LiteralTerm).text.replaceAll(nonNumberRegex, ''))!),
+              blurRadius: (blurRadius is css.LiteralTerm) ? double.tryParse((blurRadius as css.LiteralTerm).text.replaceAll(nonNumberRegex, ''))! : 0.0,
           ));
         } else {
           shadow.add(Shadow(
-            offset: Offset(
-                double.tryParse((offsetX).text.replaceAll(nonNumberRegex, ''))!,
-                double.tryParse(
-                    (offsetY).text.replaceAll(nonNumberRegex, ''))!),
-            blurRadius: (blurRadius is css.LiteralTerm)
-                ? double.tryParse(
-                    (blurRadius).text.replaceAll(nonNumberRegex, ''))!
-                : 0.0,
+              offset: Offset(
+                  double.tryParse((offsetX as css.LiteralTerm).text.replaceAll(nonNumberRegex, ''))!,
+                  double.tryParse((offsetY as css.LiteralTerm).text.replaceAll(nonNumberRegex, ''))!),
+              blurRadius: (blurRadius is css.LiteralTerm) ? double.tryParse((blurRadius as css.LiteralTerm).text.replaceAll(nonNumberRegex, ''))! : 0.0,
           ));
         }
       }
@@ -1373,47 +907,26 @@ class ExpressionMapping {
     return finalShadows;
   }
 
-  static VerticalAlign expressionToVerticalAlign(css.Expression value) {
-    if (value is css.LiteralTerm) {
-      switch (value.text) {
-        case "sub":
-          return VerticalAlign.sub;
-        case "super":
-          return VerticalAlign.sup;
-        case "bottom":
-          return VerticalAlign.bottom;
-        case "top":
-          return VerticalAlign.top;
-        case "middle":
-          return VerticalAlign.middle;
-        case "baseline":
-        default:
-          return VerticalAlign.baseline;
-      }
-    }
-
-    return VerticalAlign.baseline;
-  }
-
-  static Color stringToColor(String rawText) {
-    var text = rawText.replaceFirst('#', '');
-    if (text.length == 3) {
-      text = text.replaceAllMapped(RegExp(r"[a-f]|\d", caseSensitive: false),
-          (match) => '${match.group(0)}${match.group(0)}');
-    }
+  static Color stringToColor(String _text) {
+    var text = _text.replaceFirst('#', '');
+    if (text.length == 3)
+      text = text.replaceAllMapped(
+          RegExp(r"[a-f]|\d", caseSensitive: false),
+          (match) => '${match.group(0)}${match.group(0)}'
+      );
     if (text.length > 6) {
-      text = "0x$text";
+      text = "0x" + text;
     } else {
-      text = "0xFF$text";
+      text = "0xFF" + text;
     }
-    return Color(int.parse(text));
+    return new Color(int.parse(text));
   }
 
   static Color? rgbOrRgbaToColor(String text) {
     final rgbaText = text.replaceAll(')', '').replaceAll(' ', '');
     try {
       final rgbaValues =
-          rgbaText.split(',').map((value) => double.parse(value)).toList();
+      rgbaText.split(',').map((value) => double.parse(value)).toList();
       if (rgbaValues.length == 4) {
         return Color.fromRGBO(
           rgbaValues[0].toInt(),
@@ -1439,41 +952,28 @@ class ExpressionMapping {
     final hslText = text.replaceAll(')', '').replaceAll(' ', '');
     final hslValues = hslText.split(',').toList();
     List<double?> parsedHsl = [];
-    for (var element in hslValues) {
-      if (element.contains("%") &&
-          double.tryParse(element.replaceAll("%", "")) != null) {
+    hslValues.forEach((element) {
+      if (element.contains("%") && double.tryParse(element.replaceAll("%", "")) != null) {
         parsedHsl.add(double.tryParse(element.replaceAll("%", ""))! * 0.01);
       } else {
-        if (element != hslValues.first &&
-            (double.tryParse(element) == null ||
-                double.tryParse(element)! > 1)) {
+        if (element != hslValues.first && (double.tryParse(element) == null || double.tryParse(element)! > 1)) {
           parsedHsl.add(null);
         } else {
           parsedHsl.add(double.tryParse(element));
         }
       }
-    }
+    });
     if (parsedHsl.length == 4 && !parsedHsl.contains(null)) {
-      return HSLColor.fromAHSL(
-              parsedHsl.last!, parsedHsl.first!, parsedHsl[1]!, parsedHsl[2]!)
-          .toColor();
+      return HSLColor.fromAHSL(parsedHsl.last!, parsedHsl.first!, parsedHsl[1]!, parsedHsl[2]!).toColor();
     } else if (parsedHsl.length == 3 && !parsedHsl.contains(null)) {
-      return HSLColor.fromAHSL(
-              1.0, parsedHsl.first!, parsedHsl[1]!, parsedHsl.last!)
-          .toColor();
-    } else {
-      return Colors.black;
-    }
+      return HSLColor.fromAHSL(1.0, parsedHsl.first!, parsedHsl[1]!, parsedHsl.last!).toColor();
+    } else return Colors.black;
   }
 
   static Color? namedColorToColor(String text) {
-    String namedColor = namedColors.keys.firstWhere(
-        (element) => element.toLowerCase() == text.toLowerCase(),
-        orElse: () => "");
-    if (namedColor != "") {
-      return stringToColor(namedColors[namedColor]!);
-    } else {
-      return null;
-    }
+     String namedColor = namedColors.keys.firstWhere((element) => element.toLowerCase() == text.toLowerCase(), orElse: () => "");
+     if (namedColor != "") {
+       return stringToColor(namedColors[namedColor]!);
+     } else return null;
   }
 }
