@@ -1,5 +1,11 @@
+import 'dart:convert';
+import 'dart:math';
+
+import 'package:chewie/chewie.dart';
+import 'package:chewie_audio/chewie_audio.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_html/src/style.dart';
+import 'package:video_player/video_player.dart';
+import 'package:flutter_html/style.dart';
 
 Map<String, String> namedColors = {
   "White": "#FFFFFF",
@@ -20,6 +26,23 @@ Map<String, String> namedColors = {
   "Purple": "#800080",
 };
 
+Map<String, String> mathML2Tex = {
+  "sin": r"\sin",
+  "sinh": r"\sinh",
+  "csc": r"\csc",
+  "csch": r"csch",
+  "cos": r"\cos",
+  "cosh": r"\cosh",
+  "sec": r"\sec",
+  "sech": r"\sech",
+  "tan": r"\tan",
+  "tanh": r"\tanh",
+  "cot": r"\cot",
+  "coth": r"\coth",
+  "log": r"\log",
+  "ln": r"\ln",
+};
+
 class Context<T> {
   T data;
 
@@ -38,8 +61,7 @@ class MultipleTapGestureDetector extends InheritedWidget {
   }) : super(key: key, child: child);
 
   static MultipleTapGestureDetector? of(BuildContext context) {
-    return context
-        .dependOnInheritedWidgetOfExactType<MultipleTapGestureDetector>();
+    return context.dependOnInheritedWidgetOfExactType<MultipleTapGestureDetector>();
   }
 
   @override
@@ -56,6 +78,21 @@ class CustomBorderSide {
   Color? color;
   double width;
   BorderStyle style;
+}
+
+/// Helps keep track of controllers used by [Html] widgets.
+/// A map is used so that controllers are not duplicated on widget rebuild
+class InternalControllers {
+  Map<int, ChewieAudioController> chewieAudioControllers = {};
+  Map<int, ChewieController> chewieControllers = {};
+  Map<int, VideoPlayerController> audioPlayerControllers = {};
+  Map<int, VideoPlayerController> videoPlayerControllers = {};
+}
+
+String getRandString(int len) {
+  var random = Random.secure();
+  var values = List<int>.generate(len, (i) =>  random.nextInt(255));
+  return base64UrlEncode(values);
 }
 
 extension TextTransformUtil on String? {
@@ -86,4 +123,8 @@ extension TextTransformUtil on String? {
       return this;
     }
   }
+}
+
+extension ClampedEdgeInsets on EdgeInsetsGeometry {
+  EdgeInsetsGeometry get nonNegative => this.clamp(EdgeInsets.zero, const EdgeInsets.all(double.infinity));
 }
