@@ -2,18 +2,19 @@ import 'package:flutter/material.dart';
 import 'package:flutter_html/flutter_html.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'test_utils.dart';
+import 'test_data.dart';
 
 class TestApp extends StatelessWidget {
   final Widget body;
 
-  const TestApp(this.body, {super.key});
+  TestApp(this.body);
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       home: Scaffold(
         body: body,
+        appBar: AppBar(title: Text('flutter_html')),
       ),
     );
   }
@@ -59,18 +60,6 @@ void main() {
           """),
       ),
     );
-    //    await expectLater(find.byType(Html), matchesGoldenFile('./goldens/whitespace.png'));
-  });
-
-  testWidgets('whitespace between inline elements golden test',
-      (WidgetTester tester) async {
-    await tester.pumpWidget(
-      TestApp(
-        Html(
-          data: """<b>Harry</b> <b>Potter</b>""",
-        ),
-      ),
-    );
-    // await expectLater(find.byType(Html), matchesGoldenFile('./goldens/whitespace_btwn_inline.png'));
+//    await expectLater(find.byType(Html), matchesGoldenFile('./goldens/whitespace.png'));
   });
 }
